xwayland: Don't move window in response to geometry change if state changed
When we leave fullscreen or maximized mode we restore a saved window
position. This is expected, but that saved position was saved when window
geometry was set to have shadows rendered.

Since we restore a saved position that had shadow geometry, we don't want
to move the window when we set geometry again, or we'll move away from the
intended saved position.

I guess this is a counter-proposal to !614

Fixes #454

Signed-off-by: Derek Foreman <derek.foreman@collabora.com>